backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.1
pydantic==2.10.4
